+++ backend/capellacollab/projects/toolmodels/modelsources/git/gitlab/models.py
@@ -1,0 +1,11 @@
+# SPDX-FileCopyrightText: Copyright DB Netz AG and the capella-collab-manager contributors
+# SPDX-License-Identifier: Apache-2.0
+
+import enum
+
+
+class ModelArtifactStatus(enum.Enum):
+    SUCCESS = "success"
+    FAILURE = "failure"
+    UNCONFIGURED = "unconfigured"
+    UNSUPPORTED = "unsupported"